refactor: districts
Rename scopes to `districts` and define districts to be an ordered set
of layers.

Refine the sematics of how link processing will work on the basis of a
link having two parts - the solution identifier and the occurrence
identifier. Each of these will be identified by attributes in the link.